requirements.txt
python-dotenv
langchain
langchain_core
langchain_community
langchain-ollama
langchain_text_splitters
faiss-cpu
langchain-ollama
langchain-unstructured
unstructured-client
unstructured
"unstructured[all-docs]"
python-magic-bin
flask